+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.4.3 (Next)
 - Prevent import to existing state address in `terraform_import` module.
 - Validate `faas` function existence in `remove` module.
+- Add `log_level` parameter to `goss` modules.
 
 ### 1.4.2
 - Add new parameters for remaining `faas` plugin modules.